Update website for gem5 v24.1 (#168)
This PR updates the documentation on the gem5 website for gem5 version
24.1, changes commands/instructions to use the ALL build where possible,
and adds instructions for running gem5 with a pre-built binary in a few
places.

The documentation updates to the gem5 website are as follows:

- Updating `_pages/documentation/gem5-stdlib` parts 0-4. I tested the
scripts in the tutorials and made changes if they didn't work with gem5
v24.1.
- Updating Learning gem5 part 1, parts 1, 2, 3, and 5. The slides for
1-Introduction/02-getting-started and 2-Using-gem5/01-stdlib in the 2024
gem5 bootcamp were used to update part1/part2 and part1/part3,
respectively. The materials were modified to use fewer bullet
points/more prose, and the description of the X86DemoBoard in
1-Introduction/02-getting-started was updated for v24.1.